reorganize init widgets
