+++ test/widget_test.dart
@@ -6,10 +6,10 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:book_adapter/src/app.dart';
+import 'package:book_adapter/src/features/initialization/presentation/init_widget.dart';
 import 'package:book_adapter/src/features/library/presentation/library_view.dart';
 import 'package:book_adapter/src/service/firebase_service.dart';
 import 'package:book_adapter/src/shared/controller/firebase_controller.dart';
-import 'package:book_adapter/src/shared/widgets/init_widget.dart';
 import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_auth_mocks/firebase_auth_mocks.dart';
